refactor: address PR code review feedback
Must-fix changes:
- Extract validateFileMetadata() to shared utility (DRY principle)
- Add TODO comment for visitor ID validation (requires schema changes)
- Move test utilities to tests/utilities/ directory
- Add file upload configuration to .env.example

Changes made:
- Created src/utils/fileValidation.js with validateFileMetadata()
- Updated conversationController.js to use shared utility
- Updated agentController.js to use shared utility
- Moved 5 test files to proper tests/utilities/ directory
- Added MAX_FILE_UPLOAD_SIZE_MB and UPLOADS_DIR to .env.example
- Added detailed TODO for visitor ID validation implementation

This addresses the high-priority items from code review.
Visitor ID validation marked for follow-up PR (requires database schema changes).
