+++ custom-widget/backend/.env.example
@@ -30,6 +30,10 @@ ENABLE_AUTO_CATEGORIZATION=true
 AUTO_CATEGORIZATION_IDLE_MINUTES=15
 AUTO_CATEGORIZATION_MIN_MESSAGES=3
 
+# File Upload Configuration
+MAX_FILE_UPLOAD_SIZE_MB=10
+UPLOADS_DIR=./uploads
+
 # Mistral Configuration for Embeddings
 MISTRAL_API_KEY=your-mistral-api-key
