+++ custom-widget/backend/tests/utilities/test-categorization.js
