+++ custom-widget/backend/tests/utilities/test-clear-category.js
